refactor and rework entryscript
