--- entrypoint
@@ -1,54 +1,0 @@
-#!/usr/bin/env bash
-
-#set -m
-
-#export LC_MEASUREMENT=nb_NO.UTF-8
-
-LOG_FILE=/tmp/com.prusa3d.PrusaSlicer.log
-
-APP_ARGS=("$@")
-APP=/app/bin/prusa-slicer
-WATCH_ERROR="An error occured while setting up locale"
-LOCALE_ERROR_DETECTED=false
-
-function startup {
-    if [ $LOCALE_ERROR_DETECTED == false ]; then
-        echo "EXECUTED: $APP ${APP_ARGS[@]}"
-        echo "PRUSA_SLICER_DARK_THEME: $PRUSA_SLICER_DARK_THEME"
-    fi
-    [ $LOCALE_ERROR_DETECTED == true ] &&
-        export LC_ALL=en_US.UTF-8 && echo -e "LOCALE_ERROR_DETECTED: $LOCALE_ERROR_DETECTED"
-    [ $PRUSA_SLICER_DARK_THEME == true ] && export GTK_THEME='Adwaita:dark'
-    ($APP ${APP_ARGS[@]} &) 2>$LOG_FILE 1>/dev/null &
-    [ $PRUSA_SLICER_DARK_THEME == true ] && (/app/bin/set-dark-theme-variant.py &)
-}
-
-function monitor_app_log {
-    (tail -F -n0 $LOG_FILE &) | grep -q "$WATCH_ERROR" && LOCALE_ERROR_DETECTED=true
-    wait_period=0
-    while true; do
-
-        # if [ pgrep -f "$APP" ]; then
-        #     echo "APP_RUNNING"
-        #     sleep 1
-        # else
-        #     echo "APP_NOT_RUNNING"
-        # fi
-
-        if [ $LOCALE_ERROR_DETECTED == true ]; then
-            startup
-            LOCALE_ERROR_DETECTED=false
-            break
-        fi
-        [ $wait_period -gt 8 ] && break || sleep 1
-        wait_period=$((wait_period + 1))
-    done
-    exit 0
-}
-
-function main {
-    startup
-    monitor_app_log
-}
-
-main
